refactor: isolate OneRec graph and configure core split
