+++ xllm/core/common/help_formatter.h
@@ -116,6 +116,7 @@ const OptionCategory kRecOptions = {"REC OPTIONS",
                                      "rec_worker_max_concurrency",
                                      "enable_multistream_perf_mode",
                                      "enable_onerec_multistream_core_split",
+                                     "onerec_multistream_core_ratio",
                                      "output_rec_logprobs"}};
 
 const OptionCategory kPrefixCacheOptions = {
