+++ xllm/core/runtime/acl_graph_executor_impl.h
@@ -21,9 +21,7 @@ limitations under the License.
 
 #include <cstdint>
 #include <memory>
-#include <mutex>
 #include <optional>
-#include <string>
 #include <vector>
 
 #include "core/common/macros.h"
@@ -56,7 +54,7 @@ struct TilingBufferInfo;
 
 namespace xllm::npu {
 
-class OneRecAclGraph;
+class OneRecAclGraphExecutor;
 
 // Helper class to hold persistent parameters for graph execution
 // Multiple AclGraph instances can share the same GraphPersistentParam object
@@ -336,17 +334,7 @@ class AclGraphExecutorImpl : public ExecutorImpl {
   // Persistent parameters shared across all AclGraph instances
   std::unique_ptr<GraphPersistentParam> persistent_param_;
 
-  absl::flat_hash_map<std::string, std::unique_ptr<OneRecAclGraph>>
-      onerec_graphs_;
-  // Protect this executor's graph cache. Device-wide NPUGraph generator state
-  // is protected separately by DeviceCaptureLock.
-  std::mutex onerec_graph_mutex_;
-  bool onerec_last_first_prefill_graph_ready_ = false;
-
-  ModelOutput run_onerec_graph(const torch::Tensor& tokens,
-                               const torch::Tensor& positions,
-                               std::vector<KVCache>& kv_caches,
-                               const ModelInputParams& params);
+  std::unique_ptr<OneRecAclGraphExecutor> onerec_graph_executor_;
 
   // Get bucket num_tokens for given num_tokens
   // For num_tokens < 8: use 1, 2, 4, 8
